# Optimize performance for large-scale parallelism