--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,6 +17,7 @@ all-features = true
 codspeed-criterion-compat = { workspace = true }
 rand = { workspace = true }
 wasm-bindgen-test = { workspace = true }
+rayon = { workspace = true }
 
 [target.'cfg(unix)'.dev-dependencies]
 jemallocator = "0.5.0"
@@ -59,5 +60,6 @@ ordered-float = "4.0"
 phf = "0.11"
 phf_codegen = "0.11"
 rand = "0.8"
+rayon = "1.10"
 regex = "1.0"
 wasm-bindgen-test = "0.3.0"